pyproject.toml
[tool.poetry]
name = "saleor"
version = "2.10.0"
description = "A modular, high performance e-commerce storefront built with GraphQL, Django, and ReactJS."
authors = ["Mirumee Software <hello@mirumee.com>"]
license = "BSD-3-Clause"

readme = "README.md"

homepage = "https://getsaleor.com/"
repository = "https://github.com/mirumee/saleor"
documentation = "https://docs.getsaleor.com/en/latest/"

[tool.poetry.dependencies]
python = "~3.8"
babel = "~2.8"
boto3 = "^1.12"
braintree = "~3.49.0"
celery = { version = "^4.3.0", extras = ["redis"] }
dj-database-url = "^0"
dj-email-url = "^1"
django = "^3.0.6"
django-countries = "^6.1"
django-filter = "^2.2"
django-graphql-jwt = "0.3.0" # https://github.com/mirumee/saleor/issues/4652
django-measurement = "^3.0"
django-mptt = "^0"
django-phonenumber-field = "^4.0"
django-prices = "^2.1"
django-prices-openexchangerates = "^1.0.1"
django-prices-vatlayer = "^1.1.0"
django-storages = { version = "^1.7.1", extras = [ "google" ] }
django-templated-email = "^2.3.0"
django-versatileimagefield = "^2.0"
draftjs-sanitizer = "^1.0.0"
faker = "^4.0"
freezegun = "^0"
google-cloud-storage = "^1.18.0"
google-i18n-address = "^2.3.5"
google-measurement-protocol = "^1.0"
graphene-django = "^2.9.1"
graphene-federation = "^0.1.0"
html-to-draftjs = "^1.0.1"
markdown = "^3.1.1"
maxminddb = "^1.5.4"
maxminddb-geolite2 = "^2018.701"
opentracing = "^2.3.0"
phonenumberslite = "^8.12.2"
prices = "^1.0"
psycopg2-binary = "^2.8.3"
purl = "^1.5"
python-magic-bin = {version = "^0.4.14", platform = "win32"}
razorpay = "^1.2"
requests = "^2.22"
sentry-sdk = "^0"
stripe = "^2.47.0"
text-unidecode = "^1.2"
tqdm = "^4.36"
uwsgi = {version = "^2.0.8", platform = "!=win32"}
weasyprint = ">=48"
oauthlib = "^3.0"
jaeger-client = "^4.3.0"
django-cache-url = "^3.1.2"
django-tenant-schemas = "^1.9.0"
tenant-schemas-celery = "^0.2.1"
django-ses = "^0.8.12"
pycurl = "^7.43.0"

[tool.poetry.dev-dependencies]
black = "19.10b0"
codecov = "^2.0.22"
coverage = "^5.1"
django-debug-toolbar = "^2.0"
django-debug-toolbar-request-history = "^0"
django-graphiql-debug-toolbar = "^0.1.4"
django-extensions = "^2.2.1"
flake8 = "^3.8.1"
isort = "^4.3.21"
pre-commit = "^2.3"
pycodestyle = "^2.5"
pydocstyle = "^5.0"
pylint = "^2.3.1"
pylint-celery = "^0"
pylint-django = "^2.0.15"
pylint-plugin-utils = "^0"
pytest = "^5.4.1"
pytest-cov = "^2.7.1"
pytest-django = "^3.5.1"
pytest-django-queries = "~1.1"
pytest-mock = "^3.1.0"
pytest-vcr = "^1.0.2"
pytest-xdist = "^1.32"
tox = "^3.15.0"
transifex-client = "^0"
django-stubs = "1.2"
mypy = "0.740"
boto3-type-annotations-with-docs = "^0"
moto = "^1.3.14"

[tool.black]
target_version = ['py35', 'py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |\.hg
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
'''
